\name{hexViewFile}
\alias{hexViewFile}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Specify an Example File }
\description{
  This is just a convenience function for specifying
  one of the example files contained in the hexView package.
  It is used in examples in hexView help pages.
}
\usage{
hexViewFile(filename)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{filename}{ The name of the example file. }
}
\value{
  The full path to the appropriate example file.
}
\author{ Paul Murrell }
\examples{
hexViewFile("rawTest.txt")
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ file }
